# openapi_nip_full_20251022_021407.yaml
openapi: 3.1.0
info:
  title: MEEP Clone API
  version: 3.0.0
  description: 'Patch Plan — Primeira leva (Auth, Empresas, Usuários, Eventos, Clientes,
    PDV, Ingressos, Financeiro). Padrões aplicados: RESTful, JWT bearer, paginação,
    erros tipados, idempotência em operações financeiras, webhooks como POST, versionamento
    /api/v1.

    '
servers:
- url: https://api.seudominio.com/api/v1
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Erro padrão
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
    SortParam:
      name: sort
      in: query
      schema:
        type: string
    QueryParam:
      name: q
      in: query
      schema:
        type: string
...
  schemas:
    Empresa:
      type: object
      required:
      - nome
      properties:
        id:
          type: integer
        nome:
          type: string
        cnpj:
          type: string
        status:
          type: string
          enum:
          - ativo
          - inativo
...
paths:
...
  /auth/login:
    post:
      security: []
      summary: Auth Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - senha
              properties:
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/me:
    get:
      summary: Perfil autenticado
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/ErrorResponse'
  /empresas:
    get:
      summary: Listar empresas
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      - $ref: '#/components/parameters/SortParam'
      - $ref: '#/components/parameters/QueryParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Empresa'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      summary: Criar empresa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Empresa'
      responses:
        '201':
          description: Created
        default:
          $ref: '#/components/responses/ErrorResponse'
...